e2e: Load catalog entries from the in-cluster catalog and verify they  install
